Ds_Virtual_space_micro/services/api-gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
python-dotenv
redis